本次提交是对副本管理模块中数据写入流程的彻底重构 具体细节： 1、对用户数据做缓存，然后通过异步消费的方式进行处理； 2、文件的容量大小及文件时间区间的设置现在作为Storage的属性进行设置； 3、向FID中添加了时间区间，修复时间区间动态变化导致原来的FID无法读取文件的异常； 4、向服务管理类添加了每次对服务进行注册就更新服务注册时间的功能； 5、废弃了磁盘节点中需要通过序列号管理数据的机制； 6、修复FID之中有换行符的问题。
注：
1、文件同步功能目前没有完善，还存在问题；
2、磁盘管理节点也许要进行重构，预计会采用文件模板的方式对文件进行管理。

+++ FS_Server/src/main/java/com/bonree/brfs/disknode/data/read/DataFileReader.java
@@ -13,19 +13,19 @@ public class DataFileReader {
 	
 	private DataFileReader() {}
 	
-	public static byte[] readFile(File file, int offset, int size) {
+	public static byte[] readFile(File file, long offset, int size) {
 		return readFile(file.getAbsolutePath(), offset, size);
 	}
 	
-	public static byte[] readFile(File file, int offset) {
+	public static byte[] readFile(File file, long offset) {
 		return readFile(file, offset, Integer.MAX_VALUE);
 	}
 	
-	public static byte[] readFile(String filePath, int offset) {
+	public static byte[] readFile(String filePath, long offset) {
 		return readFile(filePath, offset, Integer.MAX_VALUE);
 	}
 	
-	public static byte[] readFile(String filePath, int offset, int size) {
+	public static byte[] readFile(String filePath, long offset, int size) {
 		RandomAccessFile file = null;
 		MappedByteBuffer buffer = null;
 		try {
